change amount & balance from double to BigDecimal

+++ src/main/java/pl/pomykalskimateusz/evolution/api/game/fetch/BetModel.java
@@ -2,7 +2,9 @@ package pl.pomykalskimateusz.evolution.api.game.fetch;
 
 import pl.pomykalskimateusz.evolution.repository.bet.BetEntity;
 
-public record BetModel(Long id, double betAmount) {
+import java.math.BigDecimal;
+
+public record BetModel(Long id, BigDecimal betAmount) {
     public static BetModel of(BetEntity betEntity) {
         return new BetModel(betEntity.getId(), betEntity.getAmount());
     }
