+++ src/main/java/pl/pomykalskimateusz/evolution/api/game/fetch/FetchGameResponse.java
@@ -2,8 +2,13 @@ package pl.pomykalskimateusz.evolution.api.game.fetch;
 
 import pl.pomykalskimateusz.evolution.repository.game.GameEntity;
 
-public record FetchGameResponse(Long id, String type, double balance, double winAmount, BetModel bet) {
+import java.math.BigDecimal;
+
+public record FetchGameResponse(Long id, String type, BigDecimal balance, BigDecimal winAmount, BetModel bet) {
     public static FetchGameResponse of(GameEntity gameEntity) {
-        return new FetchGameResponse(gameEntity.getId(), gameEntity.getType(), gameEntity.getBalance(), gameEntity.getWinAmount(), BetModel.of(gameEntity.getBet()));
+        return new FetchGameResponse(
+                gameEntity.getId(), gameEntity.getType(), gameEntity.getBalance(),
+                gameEntity.getWinAmount(), BetModel.of(gameEntity.getBet())
+        );
     }
 }
