+++ src/main/java/pl/pomykalskimateusz/evolution/domain/model/UserBet.java
@@ -2,8 +2,10 @@ package pl.pomykalskimateusz.evolution.domain.model;
 
 import pl.pomykalskimateusz.evolution.domain.exception.InvalidBetAmountError;
 
+import java.math.BigDecimal;
+
 public record UserBet(long userId, Bet bet) {
-    public static UserBet of(long userId, double amount) {
+    public static UserBet of(long userId, BigDecimal amount) {
         if(Bet.isNotValid(amount)) {
             throw new InvalidBetAmountError();
         }
